+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 **Added**
 
 - Support for custom query printers. `#21`_
+- Support for custom scalars. `#22`_
 
 **Performance**
 
@@ -153,5 +154,6 @@ Invalid queries:
 .. _#32: https://github.com/Stranger6667/hypothesis-graphql/32
 .. _#30: https://github.com/Stranger6667/hypothesis-graphql/30
 .. _#26: https://github.com/Stranger6667/hypothesis-graphql/26
+.. _#22: https://github.com/Stranger6667/hypothesis-graphql/22
 .. _#21: https://github.com/Stranger6667/hypothesis-graphql/21
 .. _#14: https://github.com/Stranger6667/hypothesis-graphql/14